Chapter9_Challenge4
New game where players can travel between locations of their choosing. Players can opt to stay, move, leave the game or end the game. Map is based on a dictionary where keys are each location and values are an array of corresponding destinations for each location; therefore a new map can be used in place of the one currently in the game